Initial Import of new Bootstrap Theme
There are quite a large number of css and js plugins that the theme
included that seem unnecessary, but might provide good reference in the
future. So I am checking in the base theme with minimal changes. The
idea is that we can have a suite of components and plugins that we might
use for building out the different Radius apps.

Main changes include:

- renaming bower_components to components
- create base angle.js and angle.css that can be included in downstream
  apps
- bump the version number
- manually moving the fonts to the fonts and stylesheets folders (used
  erb to re-write the .css files for the proper asset paths
- Stole the `fa_icon` helper from https://github.com/bokmann/font-awesome-rails