Update datetimepicker component to v4.17.47
The datetimepicker autocomplete was failing if you added
`moment-timezone-with-data` to donkeykong's application.js file.  It
would clear the field if you enter a date (e.g., 10/11) and press enter,
it it supposed to parse that date and update the field with the full
date.

This problem appears to be resolved by updating to the latest version of
datetimepicker, not sure why it was broken but if this fixes it then
that's good enough for me.