Add topnavbar to theme assets
Converted the topnavbar to erb and added the partial.

This allow us to easily update the top navigation across any site that
uses the radius theme and have it be consistent. The instigator here was
removing the link to the plans page. Ya know, since there is no more
plans page.

I added helpers for building the urls, most of them check to see if the
equivalent route is defined and return that if so. These should only be
defined in the Kracken Rails app itself, so the links in that dropdown
will use the local app.

This works great for most links. The `destroy_user_session_path` is only
defined in Kracken Rails, so can check if that exists; if it does use it, else
build the url from the `kracken_url`. It doesn't work for `teams_path`,
many of the client apps have a team controller so checking for that
won't work well.

Finding a way to detect if we are in the Kracken Rails app proved to be
tricky and out of the scope of this change.