Add custom templates :tada:
:tada:

Now we can use our Radius custom templates for a scaffolded controller
and action views.  A project using this will still need to have
`slim-rails` in it's own Gemfile since we are not monkey-patching any
generators.  However, because of that, we need to work around how the
generators work. By default, rails will look for a template, and
generate that first one it finds.  Since our custom templates are in a
gem, rails will find a template (default) in the project and spit it
out. We want to check `radius-rails` first for any templates before
going back to the defaults.  Those changes are in `lib/radius-rails.rb`.