Adds universal ValidatingEnum support for apps using this gem
Closes #32

This repackages and re-uses the ValidatingEnum code that we've tested
out on both Iris and Kracken. I wasn't completely sure on where to put
certain files in a Gem context, so some of those decisions can probably
be re-thought.

the ValidatingEnum module was put under app/models/concerns because
that's where it was on both apps. It's just a plain module though that
is extended by an AR model, so it could realistically be stored
anywhere.

ValidatingEnumType is put under lib/types for the same reason -- that's
where it was in the apps. This could also be stored anywhere. This file
was required in the lib/radius-rails.rb file. I had looked up whether it
would be preferable to load it through a RailTie, but since the original
file wasn't specifically loaded through any AR config options, just a
plain require, this seemed ok?

The one thing I'm unclear about and can definitely roll back if this is
too aggressive, is the universal extension of the module. Currently, we
have "extend ValidatingEnum" for classes that want to use it. I _did_
add a RailTie initializer here that will automatically extend it on
every ActiveRecord model. Considering we're only currently using it on a
handful of models, this might be unnecessary, though I do like that it
makes it feel more like a generalized behavior that we're adding to the
app. If the team preference is to allow this behavior only when
opted-in, I'm happy to defer to that.