Makes this gem compatible with Zeitwerk.
Zeitwerk is Rails 6's autoloading magician. It has one demand: if a file
exists in the autoloading paths, that file has to define an
appropriately-named constant.

Because of reasons, we had to disable the Validating Enum a while back.
(Ironically, because of Rails 6) The file still exists, but wasn't
defining the module.

This commit re-defines the module, but leaves it empty. This is a
harmless change, and I've verified locally that it satisfies Zeitwerk.
We don't have to also do this with the stuff under lib/ because that is
not autoloaded by default.