Updates FontAwesome to version 6
closes #46

The TLDR is that this updates our FontAwesome v4 to FontAwesomePro v6,
via Runtriz. (Thanks for the assist, Alonso!)

The full methodology is detailed in Issue #46, but the short version is
that I:

1) Identify all repos we have using radius-rails
2) Within those repos, grep for all fa-* class uses
3) Distill that down to a unique, sorted list
4) Determine which FA sets are needed to satisfy these classes

This gave me a trivial list of files that were being used, but I
suspected we could pare it down further. For each file, I reviewed
whether or not the enabling class (eg. fad / fa-duotone) was being used
anywhere in any of our apps that use radius-rails. Most of the
collections were _not_ being used, so I could omit them.

In the end we only had to include the basic set, fa-regular, fa-brands, and fa-solid.

I also incorporated the v4-shims fontface and styles to give us
backwards compatibility - there were a couple styles (fa-empire, used on
"New Order" nav link) that have been changed since v4.

The old font-awesome files have been deleted. Note that there are fewer
variations now than there used to be (just WOFF2 and TTF).

Additionally, as this is significant version bump in one of the major
dependencies, I did a major version bump for the gem 3.x -> 4.0.